refactor: moved cc & bcc button from header to 'TO' section
